Update board type
