+++ src/types/board.ts
@@ -1,0 +1,11 @@
+import type { ReverseCategoryType, CommentType, ProfileType } from '@/types';
+
+export interface PostType {
+  id: string;
+  title: string;
+  content: string;
+  createdAt: Date;
+  boardCategory: ReverseCategoryType;
+  comments: CommentType[];
+  author: ProfileType;
+}
