+++ src/types/index.ts
@@ -1,6 +1,7 @@
 export * from './auth';
 export * from './cardInfo';
 export * from './career';
+export * from './board';
 export * from './category';
 export * from './comment';
 export * from './form';
@@ -11,6 +12,5 @@ export * from './mentor';
 export * from './mentorInfo';
 export * from './opponentInfo';
 export * from './position';
-export * from './post';
 export * from './profile';
 export * from './worker';
